refactor(service): 쿠키 인증 관련 서비스 분리 및 메서드 변경
