+++ backend/src/main/java/codezap/member/service/MemberService.java
@@ -1,22 +1,16 @@
 package codezap.member.service;
 
-import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
-import java.util.Base64;
-import java.util.Objects;
-
 import jakarta.servlet.http.Cookie;
 
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import codezap.global.exception.CodeZapException;
 import codezap.member.domain.Member;
-import codezap.member.repository.MemberRepository;
 import codezap.member.dto.LoginRequest;
 import codezap.member.dto.MemberDto;
 import codezap.member.dto.SignupRequest;
+import codezap.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -24,6 +18,7 @@ import lombok.RequiredArgsConstructor;
 public class MemberService {
 
     private final MemberRepository memberRepository;
+    private final AuthService authService;
 
     public Member signup(SignupRequest request) {
         assertUniqueEmail(request.email());
@@ -33,37 +28,11 @@ public class MemberService {
     }
 
     public MemberDto login(LoginRequest request) {
-        return login(request.email(), request.password());
+        return authService.authorizeByEmailAndPassword(request.email(), request.password());
     }
 
-    public MemberDto login(Cookie[] cookies) {
-        String authHeaderValue = getAuthCookieValue(cookies);
-        String[] credentials = decodeCredentials(authHeaderValue);
-        String email = credentials[0];
-        String password = credentials[1];
-        return login(email, password);
-    }
-
-    private MemberDto login(String email, String password) {
-        Member member = memberRepository.findByEmail(email).orElseThrow(this::throwUnauthorized);
-        if (!member.matchPassword(password)) {
-            throwUnauthorized();
-        }
-        return MemberDto.from(member);
-    }
-
-    private String getAuthCookieValue(Cookie[] cookies) {
-        return Arrays.stream(cookies)
-                .filter(cookie -> Objects.equals(cookie.getName(), HttpHeaders.AUTHORIZATION))
-                .findFirst()
-                .map(Cookie::getValue)
-                .orElseThrow(this::throwUnauthorized);
-    }
-
-    private String[] decodeCredentials(String encodedCredentials) {
-        byte[] decodedBytes = Base64.getDecoder().decode(encodedCredentials.getBytes(StandardCharsets.UTF_8));
-        String decodedString = new String(decodedBytes);
-        return decodedString.split(":");
+    public MemberDto checkLogin(Cookie[] cookies) {
+        return authService.authorizeByCookie(cookies);
     }
 
     private void assertUniqueEmail(String email) {
@@ -84,9 +53,5 @@ public class MemberService {
 
     public boolean isUniqueUsername(String username) {
         return !memberRepository.existsByUsername(username);
-    }
-
-    private CodeZapException throwUnauthorized() {
-        throw new CodeZapException(HttpStatus.UNAUTHORIZED, "인증에 실패했습니다.");
     }
 }
